--Moved non-metadata specific abstract classes/class tempatles to core
AbstractManagedObject defines a set of methods that are independent of application.  Any object that inherits from AbstractManagedObject can be managed by a ManagedContainer, so these constructions were moved from Metadata directory to core to be more easily accessed everywhere; AbsObjAttributesManager class template was added to provide functionality that requires the managed objects be instances of AbstractObjectAttributes (currently stage and object);
