+++ src/esp/core/CMakeLists.txt
@@ -30,12 +30,14 @@ find_package(Corrade REQUIRED Utility)
 
 add_library(
   core STATIC
+  AbstractManagedObject.h
   Buffer.cpp
   Buffer.h
   Configuration.h
   esp.cpp
   esp.h
   logging.h
+  ManagedContainerBase.h
   random.h
   spimpl.h
   Utility.h
