+++ src/esp/metadata/managers/ObjectAttributesManager.cpp
@@ -3,7 +3,7 @@
 // LICENSE file in the root directory of this source tree.
 
 #include "ObjectAttributesManager.h"
-#include "AttributesManagerBase.h"
+#include "AbsObjAttributesManagerBase.h"
 
 #include <Corrade/Utility/String.h>
